+++ data/__init__.py
@@ -1,4 +1,4 @@
-VERSION = 0.005
+VERSION = 0.006
 
 NAME = f'VZero'