+++ logic/mods/dev/items/shop_old.py
@@ -10,6 +10,21 @@ ITEM_STATS = import_('items.items').ITEM_STATS
 
 class Shop:
     @classmethod
+    def buy_item(cls, api, uid, item):
+        item = cls.find_iid(item)
+        assert item is not None
+
+        if not cls.check_cost(api, uid, item):
+            logger.debug(f'{api.units[uid].name} missing gold for: {item.name} ({item.value})')
+            return FAIL_RESULT.MISSING_COST
+
+        result = cls._do_buy(api, uid, item)
+
+        # api.add_visual_effect(VisualEffect.SFX, 5, params={'sfx': 'shop'})
+        logger.debug(f'{api.units[uid].name} bought item: {item.name} ({item.value})')
+        return result
+
+    @classmethod
     def find_iid(cls, iid):
         for item in ITEM:
             if item.value == round(iid):
@@ -23,7 +38,7 @@ class Shop:
         if item is None:
             return FAIL_RESULT.MISSING_TARGET
         if not cls.check_cost(api, uid, item):
-            logger.debug(f'{api.units[uid].name} missing gold for : {item.name} ({item.value})')
+            logger.debug(f'{api.units[uid].name} missing gold for: {item.name} ({item.value})')
             return FAIL_RESULT.MISSING_COST
         result = cls._do_buy(api, uid, item)
 
@@ -38,14 +53,14 @@ class Shop:
 
     @classmethod
     def check_cost(cls, api, uid, item):
-        return api.get_stats(uid, STAT.GOLD) >= ITEM_STATS[item]['cost']
+        return api.get_stats(uid, STAT.GOLD) >= ITEM_STATS[item].cost
 
     @classmethod
     def apply_cost(cls, api, uid, item):
-        api.set_stats(uid, STAT.GOLD, -ITEM_STATS[item]['cost'], additive=True)
+        api.set_stats(uid, STAT.GOLD, -ITEM_STATS[item].cost, additive=True)
 
     @classmethod
     def apply_stats(cls, api, uid, item):
-        for stat, values in ITEM_STATS[item]['stats'].items():
+        for stat, values in ITEM_STATS[item].stats.items():
             for value_name, value in values.items():
                 api.set_stats(uid, stat, value, value_name=value_name, additive=True)